+++ src/index.ts
@@ -1,5 +1,5 @@
 export * from './extensions';
-export * from './login';
+export * from './LoginUtils';
 export * from './LoginMenu';
 export {default} from './phovea';
 export * from './watcher';
